fix: move extended-onboarding and onboarding-carousel inside onboarding
