+++ src/modules/onboarding/extended-onboarding/ExtendedOnboarding_AlsoGoodToKnowScreen.tsx
