+++ src/modules/onboarding/extended-onboarding/ExtendedOnboarding_GoodToKnowScreen.tsx
@@ -4,7 +4,7 @@ import {ExtendedOnboardingTexts, useTranslation} from '@atb/translations';
 import React from 'react';
 import {extendedOnboardingId} from '.';
 import {ArrowRight} from '@atb/assets/svg/mono-icons/navigation';
-import {OnboardingScreenComponent} from '../onboarding';
+import {OnboardingScreenComponent} from '..';
 
 export const ExtendedOnboarding_GoodToKnowScreen = () => {
   const {t} = useTranslation();
