+++ src/modules/onboarding/index.ts
@@ -18,3 +18,6 @@ export type {
   OnboardingSectionConfig,
   ShouldShowArgsType,
 } from './types';
+
+export * from './extended-onboarding';
+export * from './onboarding-carousel';
