+++ src/modules/onboarding/onboarding-carousel/index.ts
@@ -4,3 +4,4 @@ export type {
   OnboardingCarouselConfigId,
 } from './types';
 export {useNavigateToNextOnboardingCarouselScreen} from './use-navigate-to-next-onboarding-screen';
+export type {OnboardingCarouselScreenProps} from './navigation-types';
