+++ src/modules/onboarding/onboarding-carousel/onboarding-carousel-config.ts
