+++ src/modules/onboarding/onboarding-carousel/types.ts
