+++ src/modules/onboarding/onboarding-carousel/use-navigate-to-next-onboarding-screen.ts
