+++ src/modules/smart-park-and-ride/onboarding/SmartParkAndRideOnboarding_InformationScreen.tsx
@@ -5,7 +5,7 @@ import {OnboardingScreenComponent} from '@atb/modules/onboarding';
 import {ThemedCarValidTicket} from '@atb/theme/ThemedAssets';
 import {ArrowRight} from '@atb/assets/svg/mono-icons/navigation';
 import {ThemeText} from '@atb/components/text';
-import {useNavigateToNextOnboardingCarouselScreen} from '@atb/modules/onboarding-carousel';
+import {useNavigateToNextOnboardingCarouselScreen} from '@atb/modules/onboarding';
 import {sparPilotEnrollmentId} from './config';
 
 export const SmartParkAndRideOnboarding_InformationScreen = () => {
